Refactored tab bodies, allow content to decide behaviour.
The more stuff that's been added, the clearer it was becoming that a 'global' solution for tabs was becoming less practical. Some tabs have fixed sections on the left and a scrolling right, other tabs have horizontal scrolling with fixed buttons at the bottom. This commit simply changes the handling of such things to the content itself so it can be better varied.