Moved bleep volume from Mixer tab to Mic tab
Justification comment, as of Windows App version 1.4.34.205 the Bleep Volume now gets stored in the Microphone Profile instead of a 'Global' level. The GoXLR Utility was changed to support this in c6efedb.

Under the current behaviour, if you had changed the bleep volume on the mixer, and saved the profile, all the mixer volumes would have been saved but the bleep volume wouldn't have, which is counter-intuitive. Placing it on the Mic tab with the other mic settings and a relevant save button makes more logical sense.

This change also allowed me to clean up the mixer code!